feat(angular): split get webpack config in one per supported version
