+++ app/angular/src/server/angular-cli-webpack-13.x.x.d.ts
@@ -1,1 +1,7 @@
-export declare function getWebpackConfig(baseConfig: any, options: any): any;
+import { JsonObject } from '@angular-devkit/core';
+import { BuilderContext } from '@angular-devkit/architect';
+
+export declare function getWebpackConfig(
+  baseConfig: any,
+  options: { builderOptions: JsonObject; builderContext: BuilderContext }
+): any;
